player scene data API

+++ src/global/gameDB/data/player_scene_data.go
@@ -10,15 +10,15 @@ type PlayerSceneData struct {
 	Level       int64     `json:"level"`
 	Exp         int32     `json:"exp"`
 	MapId       int32     `json:"mapId"`
-	X           float32   `json:"x"`
-	Y           float32   `json:"y"`
-	Z           float32   `json:"z"`
-	DirX        float32   `json:"dirX"`
-	DirY        float32   `json:"dirY"`
-	DirZ        float32   `json:"dirZ"`
+	X           float64   `json:"x"`
+	Y           float64   `json:"y"`
+	Z           float64   `json:"z"`
+	DirX        float64   `json:"dirX"`
+	DirY        float64   `json:"dirY"`
+	DirZ        float64   `json:"dirZ"`
 	BirthMapId  int32     `json:"birthMapId"`
-	BirthX      float32   `json:"birthX"`
-	BirthY      float32   `json:"birthY"`
-	BirthZ      float32   `json:"birthZ"`
+	BirthX      float64   `json:"birthX"`
+	BirthY      float64   `json:"birthY"`
+	BirthZ      float64   `json:"birthZ"`
 	LastLoginAt time.Time `json:"lastLoginAt"`
 }
